Removing old mapper
